pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "langsmith>=0.7.11",
    "notebook>=7.5.4",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
